Update README.md
There are two tasks for the fourth homework assignment, the first is to write a Python script file to compute the Mandelbrot fractal with some of the given iteration steps and the second is to write a Python script to perform six other sub-tasks. Throughout the assignment, I defined two functions, one for each task, using double for loops to proceed the iterations that generate the plots.